core: extract simlin-core package from simlin-engine
Create a new simlin-core crate containing shared data types and utilities
that can be used by multiple crates without pulling in the full simulation
engine:

- datamodel.rs: Project, Model, Variable, SimSpecs, Dimension, etc.
- common.rs: error types, identifier types, canonicalize function
- results.rs: Results, Specs, and Method types

This allows simlin-compat to depend on simlin-core for data types without
requiring the full simlin-engine dependency. simlin-engine re-exports these
types from simlin-core for backwards compatibility. Protobuf serialization
and the simulation runtime remain in simlin-engine.

The simlin-core crate has minimal dependencies: only regex and lazy_static
for canonicalization. Neither datamodel nor common use serde - they only
use standard derives (Debug, Clone, PartialEq).
