+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "src/libsimlin",
     "src/simlin-cli",
     "src/simlin-compat",
+    "src/simlin-core",
     "src/simlin-engine",
     "src/xmutil",
 ]
