+++ src/simlin-compat/Cargo.toml
@@ -18,6 +18,7 @@ csv = { version = "1", optional = true }
 float-cmp = "0.10"
 quick-xml = { version = "0.38", features = [ "serialize", "overlapped-lists" ] }
 serde = { version = "1", features = [ "derive" ] }
+simlin-core = { version = "0.1", path = "../simlin-core" }
 simlin-engine = { version = "0.1", path = "../simlin-engine", default-features = false }
 xmutil = { version = "1", path = "../xmutil", optional = true }
