+++ src/simlin-core/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "simlin-core"
+version = "0.1.0"
+authors = ["Bobby Powers <bobbypowers@gmail.com>"]
+license = "Apache-2.0"
+edition = "2024"
+
+[lib]
+crate-type = ["rlib"]
+
+[dependencies]
+lazy_static = "1"
+regex = "1"
