+++ src/simlin-engine/Cargo.toml
@@ -15,6 +15,7 @@ schema = ["dep:schemars"]
 ai_info = ["dep:ed25519", "dep:base64"]
 
 [dependencies]
+simlin-core = { version = "0.1", path = "../simlin-core" }
 lazy_static = "1"
 regex = "1"
 unicode-xid = "0.2"
